+++ locksmith/__tests__/controllers/v2/lockSettingController.test.ts
@@ -11,7 +11,9 @@ const lockSettingMock = {
   lockAddress: '0xF3850C690BFF6c1E343D2449bBbbb00b0E934f7b',
   network,
   sendEmail: true,
+  unlockFeeChargedToUser: true,
   creditCardPrice: 0.04,
+  creditCardCurrency: 'usd',
   emailSender: 'Custom Sender',
   slug: 'slug-test',
   replyTo: 'example@gmail.com',
@@ -70,7 +72,7 @@ describe('LockSettings v2 endpoints for lock', () => {
   })
 
   it('should correctly enable "sendEmail" setting when user is lockManager', async () => {
-    expect.assertions(3)
+    expect.assertions(4)
 
     const { loginResponse } = await loginRandomUser(app)
     expect(loginResponse.status).toBe(200)
@@ -80,15 +82,17 @@ describe('LockSettings v2 endpoints for lock', () => {
       .set('authorization', `Bearer ${loginResponse.body.accessToken}`)
       .send({
         sendEmail: true,
+        unlockFeeChargedToUser: false,
       })
 
     const response = saveSettingResponse.body
     expect(saveSettingResponse.status).toBe(200)
     expect(response.sendEmail).toBe(true)
+    expect(response.unlockFeeChargedToUser).toBe(false)
   })
 
   it('should correctly save settings when user is lockManager', async () => {
-    expect.assertions(8)
+    expect.assertions(10)
 
     const { loginResponse } = await loginRandomUser(app)
     const saveSettingResponse = await request(app)
@@ -96,11 +100,13 @@ describe('LockSettings v2 endpoints for lock', () => {
       .set('authorization', `Bearer ${loginResponse.body.accessToken}`)
       .send({
         sendEmail: false,
+        unlockFeeChargedToUser: false,
         replyTo: 'example@gmail.com',
         slug: 'slug-demo',
         emailSender: 'Example',
         checkoutConfigId: 'f549d936-24e6-49e0-9acb',
         hookGuildId: '14',
+        creditCardCurrency: 'eur',
       })
 
     const response = saveSettingResponse.body
@@ -112,10 +118,12 @@ describe('LockSettings v2 endpoints for lock', () => {
     expect(response.emailSender).toBe('Example')
     expect(response.checkoutConfigId).toBe('f549d936-24e6-49e0-9acb')
     expect(response.hookGuildId).toBe(14)
+    expect(response.unlockFeeChargedToUser).toBe(false)
+    expect(response.creditCardCurrency).toBe('eur')
   })
 
   it('should save and retrieve setting when user is lockManager', async () => {
-    expect.assertions(17)
+    expect.assertions(21)
 
     const { loginResponse } = await loginRandomUser(app)
 
@@ -125,23 +133,27 @@ describe('LockSettings v2 endpoints for lock', () => {
       .set('authorization', `Bearer ${loginResponse.body.accessToken}`)
       .send({
         sendEmail: false,
+        unlockFeeChargedToUser: false,
         replyTo: 'example@gmail.com',
         creditCardPrice: 0.04,
         slug: 'slug-test-2',
         emailSender: 'Custom Sender',
         checkoutConfigId: 'f549d936-24e6-49e0-9acb',
         hookGuildId: '199',
+        creditCardCurrency: 'eur',
       })
 
     const response = saveSettingResponse.body
     expect(saveSettingResponse.status).toBe(200)
     expect(response.sendEmail).toBe(false)
+    expect(response.unlockFeeChargedToUser).toBe(false)
     expect(response.replyTo).toBe('example@gmail.com')
     expect(response.creditCardPrice).toBe(0.04)
     expect(response.slug).toBe('slug-test-2')
     expect(response.emailSender).toBe('Custom Sender')
     expect(response.checkoutConfigId).toBe('f549d936-24e6-49e0-9acb')
     expect(response.hookGuildId).toBe(199)
+    expect(response.creditCardCurrency).toBe('eur')
 
     // retrieve settings
     const getSettingResponse = await request(app)
@@ -150,6 +162,7 @@ describe('LockSettings v2 endpoints for lock', () => {
 
     expect(getSettingResponse.status).toBe(200)
     expect(getSettingResponse.body.sendEmail).toBe(false)
+    expect(getSettingResponse.body.unlockFeeChargedToUser).toBe(false)
     expect(getSettingResponse.body.replyTo).toBe('example@gmail.com')
     expect(getSettingResponse.body.lockAddress).toBe(
       lockSettingMock.lockAddress
@@ -161,10 +174,11 @@ describe('LockSettings v2 endpoints for lock', () => {
       'f549d936-24e6-49e0-9acb'
     )
     expect(getSettingResponse.body.hookGuildId).toBe(199)
+    expect(getSettingResponse.body.creditCardCurrency).toBe('eur')
   })
 
   it('should retrieve default settings for a lock', async () => {
-    expect.assertions(8)
+    expect.assertions(10)
 
     const { loginResponse } = await loginRandomUser(app)
 
@@ -176,12 +190,14 @@ describe('LockSettings v2 endpoints for lock', () => {
     expect(getSettingResponse.body.sendEmail).toBe(
       DEFAULT_LOCK_SETTINGS.sendEmail
     )
+    expect(getSettingResponse.body.unlockFeeChargedToUser).toBe(true)
     expect(getSettingResponse.body.replyTo).toBe(undefined)
     expect(getSettingResponse.body.creditCardPrice).toBe(undefined)
     expect(getSettingResponse.body.slug).toBe(undefined)
     expect(getSettingResponse.body.emailSender).toBe(undefined)
     expect(getSettingResponse.body.checkoutConfigId).toBe(undefined)
     expect(getSettingResponse.body.hookGuildId).toBe(undefined)
+    expect(getSettingResponse.body.creditCardCurrency).toBe('usd') // default currency is usd
   })
 
   it('should save null values for settings', async () => {
@@ -207,5 +223,20 @@ describe('LockSettings v2 endpoints for lock', () => {
     expect(response.emailSender).toBe(null)
     expect(response.checkoutConfigId).toBe(null)
     expect(response.hookGuildId).toBe(null)
+  })
+
+  it('should fail to save not supported currency', async () => {
+    expect.assertions(1)
+
+    const { loginResponse } = await loginRandomUser(app)
+    // save settings
+    const saveSettingResponse = await request(app)
+      .post(`/v2/lock-settings/${network}/locks/${lockSettingMock.lockAddress}`)
+      .set('authorization', `Bearer ${loginResponse.body.accessToken}`)
+      .send({
+        creditCardCurrency: 'ooo',
+      })
+
+    expect(saveSettingResponse.status).toBe(500)
   })
 })