+++ locksmith/src/operations/pricingOperations.ts
@@ -4,6 +4,7 @@ import { MIN_PAYMENT_STRIPE_CREDIT_CARD } from '../utils/constants'
 import { ethers } from 'ethers'
 import { Web3Service, getErc20Decimals } from '@unlock-protocol/unlock-js'
 import * as lockSettingOperations from './lockSettingOperations'
+import { Currencies } from '@unlock-protocol/core'
 
 interface Price {
   decimals: number
@@ -60,6 +61,14 @@ export function fromDecimal(num: string, decimals: number) {
   )
 }
 
+export function getCurrencySymbol(currency?: string) {
+  return (
+    Currencies.find(
+      (item) => item?.currency?.toLowerCase() === currency?.toLowerCase()
+    )?.symbol || '$'
+  )
+}
+
 /** Helper to return usd pricing object */
 export const toUsdPricing = ({
   amount,
@@ -91,10 +100,11 @@ export const getPricingFromSettings = async ({
   network,
   recipients = [],
 }: GetPriceProps): Promise<KeyPricingPrice | null> => {
-  const { creditCardPrice } = await lockSettingOperations.getSettings({
-    lockAddress,
-    network,
-  })
+  const { creditCardPrice, creditCardCurrency } =
+    await lockSettingOperations.getSettings({
+      lockAddress,
+      network,
+    })
 
   // return pricing object using the price from the settings
   if (creditCardPrice) {
@@ -103,10 +113,12 @@ export const getPricingFromSettings = async ({
     const amountInCents = creditCardPrice * keysToPurchase // this total is in basisPoints
     const amountInUSD = amountInCents / 100 // get total price in USD
 
+    const symbol = getCurrencySymbol(creditCardCurrency)
+
     return {
       amount: amountInUSD, // amount is usd for the single key
       decimals: 18,
-      symbol: '$',
+      symbol,
       amountInUSD,
       amountInCents,
     }
@@ -196,7 +208,7 @@ export const getTotalCharges = async ({
     }
   }
   const subtotal = Math.round(pricing.priceInAmount * 100)
-  const fees = getFees({
+  const fees = await getFees({
     subtotal,
     gasCost,
   })