+++ locksmith/src/routes/metadata.ts
@@ -1,6 +1,7 @@
 import express from 'express'
 import signatureValidationMiddleware from '../middlewares/signatureValidationMiddleware'
 import MetadataController from '../controllers/metadataController'
+import { createCacheMiddleware } from '../utils/middlewares/cacheMiddleware'
 
 const router = express.Router({ mergeParams: true })
 
@@ -71,7 +72,7 @@ router.get(
 )
 
 router.get('/:address/keyHolderMetadata', MetadataController.keyHolderMetadata)
-router.get('/:address/:keyId', MetadataController.data)
+router.get('/:address/:keyId', createCacheMiddleware(), MetadataController.data)
 router.put('/:address/:keyId', MetadataController.updateKeyMetadata)
 router.put('/:address', MetadataController.updateDefaults)
 router.put('/:address/user/:userAddress', MetadataController.updateUserMetadata)