+++ locksmith/src/websub/jobs/expiringKeys.ts
@@ -10,8 +10,8 @@ import dayjs from 'dayjs'
 import { logger } from '../../logger'
 import { hasReminderAlreadySent } from '../../operations/keyExpirationReminderOperations'
 import { sendEmail } from '../../operations/wedlocksOperations'
-import * as userMetadataOperations from './../../operations/userMetadataOperations'
-import * as membershipOperations from './../../operations/membershipOperations'
+import * as userMetadataOperations from '../../operations/userMetadataOperations'
+import * as membershipOperations from '../../operations/membershipOperations'
 import * as Normalizer from '../../utils/normalizer'
 
 /**