+++ packages/unlock-js/src/web3Service.ts
@@ -1063,4 +1063,62 @@ export default class Web3Service extends UnlockService {
       estimatedGasUsedUSD,
     }
   }
+
+  async getTokenIdsFromTx({
+    params: { hash, network, lockAddress },
+  }: {
+    params: {
+      hash: string
+      network: number
+      lockAddress: string
+    }
+  }) {
+    const provider = this.providerForNetwork(network)
+    const lockContract = await this.getLockContract(lockAddress, provider)
+    const txReceipt = await provider.getTransactionReceipt(hash)
+    const parser = lockContract.interface
+    const events = txReceipt.logs.map((log) => {
+      if (log.address.toLowerCase() !== lockAddress.toLowerCase()) return // Filter events not emitted by the lock contract
+      return parser.parseLog(log)
+    })
+
+    const purchaseItems = events.filter((event) => {
+      return event && event.name === 'Transfer'
+    })
+
+    if (purchaseItems.length) {
+      return purchaseItems.map((item) => item?.args?.tokenId?.toString())
+    }
+
+    const extendItems = events.filter((event) => {
+      return event && event.name === 'KeyExtended'
+    })
+
+    if (extendItems.length) {
+      return extendItems.map((item) => item?.args?.tokenId?.toString())
+    }
+    return null
+  }
+
+  async getGasRefundValue({
+    network,
+    lockAddress,
+  }: {
+    network: number
+    lockAddress: string
+  }) {
+    const provider = this.providerForNetwork(network)
+    const lockContract = await this.getLockContract(lockAddress, provider)
+    if (!lockContract.gasRefundValue) {
+      return '0'
+    }
+    const gasRefund = await lockContract.gasRefundValue()
+    let decimals = this.networks[network].nativeCurrency.decimals
+    const erc20Address = await lockContract.tokenAddress()
+
+    if (erc20Address !== ethers.constants.AddressZero) {
+      decimals = await getErc20Decimals(erc20Address, this.provider)
+    }
+    return ethers.utils.formatUnits(gasRefund, decimals)
+  }
 }