+++ unlock-app/src/components/interface/checkout/Connected.tsx
@@ -64,7 +64,12 @@ export function SignedIn({
 
 interface SignedOutProps {
   authenticateWithProvider(
-    provider: 'METAMASK' | 'UNLOCK' | 'WALLET_CONNECT' | 'COINBASE'
+    provider:
+      | 'METAMASK'
+      | 'UNLOCK'
+      | 'WALLET_CONNECT'
+      | 'COINBASE'
+      | 'DELEGATED_PROVIDER'
   ): Promise<void>
   onUnlockAccount(): void
   injectedProvider: any
@@ -184,7 +189,8 @@ export function Connected({
   })
   const { signIn, signOut, isSignedIn } = useSIWE()
   const [isDisconnecting, setIsDisconnecting] = useState(false)
-  const autoconnect = state.context?.paywallConfig?.autoconnect
+  const useDelegatedProvider =
+    state.context?.paywallConfig?.useDelegatedProvider
 
   useEffect(() => {
     const autoSignIn = async () => {
@@ -192,7 +198,7 @@ export function Connected({
         !isSignedIn &&
         !signing &&
         connected &&
-        (isUnlockAccount || autoconnect)
+        (isUnlockAccount || useDelegatedProvider)
       ) {
         setSigning(true)
         await signIn()
@@ -200,14 +206,14 @@ export function Connected({
       }
     }
     autoSignIn()
-  }, [connected, autoconnect, isUnlockAccount, signIn, signing, isSignedIn])
-
-  // Autoconnect
-  useEffect(() => {
-    if (autoconnect) {
-      authenticateWithProvider('METAMASK')
-    }
-  }, [autoconnect, authenticateWithProvider])
+  }, [
+    connected,
+    useDelegatedProvider,
+    isUnlockAccount,
+    signIn,
+    signing,
+    isSignedIn,
+  ])
 
   useEffect(() => {
     if (!account) {
@@ -215,7 +221,7 @@ export function Connected({
     } else console.debug(`Connected as ${account}`)
   }, [account])
 
-  if (autoconnect) {
+  if (useDelegatedProvider) {
     return <div className="space-y-2">{children}</div>
   }