+++ unlock-app/src/components/interface/checkout/main/Confirm/ConfirmCard.tsx
@@ -23,6 +23,7 @@ import { formatNumber } from '~/utils/formatter'
 import { formatFiatPriceFromCents } from '../utils'
 import { useGetTotalCharges } from '~/hooks/usePrice'
 import { useGetLockSettings } from '~/hooks/useLockSettings'
+import { getCurrencySymbol } from '~/utils/currency'
 
 interface Props {
   injectedProvider: unknown
@@ -37,6 +38,8 @@ interface CreditCardPricingBreakdownProps {
   unlockServiceFee: number
   gasCosts?: number
   loading?: boolean
+  symbol?: string
+  unlockFeeChargedToUser?: boolean
 }
 
 export function CreditCardPricingBreakdown({
@@ -45,6 +48,8 @@ export function CreditCardPricingBreakdown({
   creditCardProcessingFee,
   gasCosts,
   loading,
+  symbol = '$',
+  unlockFeeChargedToUser = true,
 }: CreditCardPricingBreakdownProps) {
   return (
     <div className="flex flex-col gap-2 pt-4 text-sm">
@@ -60,18 +65,20 @@ export function CreditCardPricingBreakdown({
         </a>
       </h3>
       <div className="divide-y">
-        <Detail
-          loading={loading}
-          className="flex justify-between w-full py-2 text-sm border-t border-gray-300"
-          label="Service Fee"
-          labelSize="tiny"
-          valueSize="tiny"
-          inline
-        >
-          <div className="font-normal">
-            {formatFiatPriceFromCents(unlockServiceFee)}
-          </div>
-        </Detail>
+        {unlockFeeChargedToUser && !loading && (
+          <Detail
+            loading={loading}
+            className="flex justify-between w-full py-2 text-sm border-t border-gray-300"
+            label="Service Fee"
+            labelSize="tiny"
+            valueSize="tiny"
+            inline
+          >
+            <div className="font-normal">
+              {formatFiatPriceFromCents(unlockServiceFee, symbol)}
+            </div>
+          </Detail>
+        )}
         {!!creditCardProcessingFee && (
           <Detail
             loading={loading}
@@ -82,7 +89,7 @@ export function CreditCardPricingBreakdown({
             inline
           >
             <div className="font-normal">
-              {formatFiatPriceFromCents(creditCardProcessingFee)}
+              {formatFiatPriceFromCents(creditCardProcessingFee, symbol)}
             </div>
           </Detail>
         )}
@@ -96,7 +103,7 @@ export function CreditCardPricingBreakdown({
             inline
           >
             <div className="font-normal">
-              {formatFiatPriceFromCents(gasCosts)}
+              {formatFiatPriceFromCents(gasCosts, symbol)}
             </div>
           </Detail>
         )}
@@ -108,7 +115,9 @@ export function CreditCardPricingBreakdown({
           valueSize="tiny"
           inline
         >
-          <div className="font-normal">{formatFiatPriceFromCents(total)}</div>
+          <div className="font-normal">
+            {formatFiatPriceFromCents(total, symbol)}
+          </div>
         </Detail>
       </div>
     </div>
@@ -175,10 +184,11 @@ export function ConfirmCard({
       config.networks[lock!.network].nativeCurrency.symbol
     ),
   })
-  const { data: { creditCardPrice } = {} } = useGetLockSettings({
-    network: lock!.network,
-    lockAddress: lock!.address,
-  })
+  const { data: { creditCardPrice, unlockFeeChargedToUser } = {} } =
+    useGetLockSettings({
+      network: lock!.network,
+      lockAddress: lock!.address,
+    })
 
   const isPricingDataAvailable =
     !isPricingDataLoading && !isPricingDataError && !!pricingData
@@ -271,6 +281,15 @@ export function ConfirmCard({
     setIsConfirming(false)
   }
 
+  const { data: { creditCardCurrency = 'usd ' } = {} } = useGetLockSettings({
+    lockAddress,
+    network: lock!.network,
+  })
+
+  const creditCardCurrencySymbol = getCurrencySymbol(creditCardCurrency)
+
+  const isError = isPricingDataError
+
   return (
     <Fragment>
       <main className="h-full p-6 space-y-2 overflow-auto">
@@ -279,7 +298,8 @@ export function ConfirmCard({
             <h4 className="text-xl font-bold"> {lock!.name}</h4>
             <ViewContract lockAddress={lock!.address} network={lockNetwork} />
           </div>
-          {isPricingDataError && (
+
+          {isError && (
             // TODO: use actual error from simulation
             <div>
               <p className="text-sm font-bold">
@@ -288,6 +308,8 @@ export function ConfirmCard({
               </p>
             </div>
           )}
+
+          {/* Breakdown of each keys */}
           {!isLoading && isPricingDataAvailable && (
             <PricingData
               network={lockNetwork}
@@ -297,43 +319,48 @@ export function ConfirmCard({
             />
           )}
         </div>
-        {!isPricingDataAvailable && (
-          <div>
-            {isLoading ? (
-              <div className="flex flex-col items-center gap-2">
-                {recipients.map((user) => (
-                  <div
-                    key={user}
-                    className="w-full p-4 bg-gray-100 rounded-lg animate-pulse"
-                  />
-                ))}
-              </div>
-            ) : (
-              <Pricing
-                keyPrice={
-                  pricingData!.total <= 0
-                    ? 'FREE'
-                    : `${formatNumber(
-                        pricingData!.total
-                      ).toLocaleString()} ${symbol}`
-                }
-                usdPrice={
-                  totalPricing?.total
-                    ? `~${formatNumber(totalPricing?.total).toLocaleString()}`
-                    : ''
-                }
-                isCardEnabled={!!creditCardEnabled}
+
+        {/* Totals */}
+        {isLoading && (
+          <div className="flex flex-col items-center gap-2">
+            {recipients.map((user) => (
+              <div
+                key={user}
+                className="w-full p-4 bg-gray-100 rounded-lg animate-pulse"
               />
-            )}
+            ))}
           </div>
         )}
-        <CreditCardPricingBreakdown
-          loading={isTotalPricingDataLoading || !isTotalPricingDataFetched}
-          total={totalPricing?.total ?? 0}
-          creditCardProcessingFee={totalPricing?.creditCardProcessingFee}
-          unlockServiceFee={totalPricing?.unlockServiceFee ?? 0}
-          gasCosts={gasCosts}
-        />
+
+        {pricingData && !isError && (
+          <Pricing
+            keyPrice={
+              pricingData.total <= 0
+                ? 'FREE'
+                : `${formatNumber(
+                    pricingData.total
+                  ).toLocaleString()} ${symbol}`
+            }
+            usdPrice={
+              totalPricing?.total
+                ? `~${formatNumber(totalPricing?.total).toLocaleString()}`
+                : ''
+            }
+            isCardEnabled={!!creditCardEnabled}
+          />
+        )}
+
+        {!isError && pricingData && (
+          <CreditCardPricingBreakdown
+            loading={isTotalPricingDataLoading || !isTotalPricingDataFetched}
+            total={totalPricing?.total ?? 0}
+            creditCardProcessingFee={totalPricing?.creditCardProcessingFee}
+            unlockServiceFee={totalPricing?.unlockServiceFee ?? 0}
+            gasCosts={gasCosts}
+            symbol={creditCardCurrencySymbol}
+            unlockFeeChargedToUser={unlockFeeChargedToUser}
+          />
+        )}
       </main>
       <footer className="grid items-center px-6 pt-6 border-t">
         <Connected
@@ -343,7 +370,7 @@ export function ConfirmCard({
           <div className="grid">
             <Button
               loading={isConfirming}
-              disabled={isConfirming || isLoading}
+              disabled={isConfirming || isLoading || isError}
               onClick={async (event) => {
                 event.preventDefault()
                 if (metadata) {