+++ unlock-protocol-com/src/config/unlock.ts
@@ -4,8 +4,8 @@ type Config = Record<
 >
 
 const devConfig: Config = {
-  gaId: process.env.NEXT_PUBLIC_UNLOCK_GA_ID ?? '0',
-  gaTmId: process.env.NEXT_PUBLIC_UNLOCK_GA_TM_ID ?? '0',
+  gaId: '',
+  gaTmId: '',
   baseURL:
     process.env.NEXT_PUBLIC_URL_BASE ?? 'https://staging.unlock-protocol.com',
   appURL:
@@ -15,12 +15,10 @@ const devConfig: Config = {
 }
 
 const stagingConfig: Config = {
-  gaId: '0',
-  gaTmId: '0',
+  gaId: '',
+  gaTmId: '',
   baseURL: 'https://staging.unlock-protocol.com',
-  appURL:
-    process.env.NEXT_PUBLIC_UNLOCK_APP_URI ??
-    'https://staging-app.unlock-protocol.com',
+  appURL: 'https://staging-app.unlock-protocol.com',
   gApiKey: 'AIzaSyBqLebWxCpOw_HO4k0KhYkWhkrS__O3XME',
 }