+++ packages/unlock-js/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changes
 
+# 0.39.0
+
+- Add support for setGasRefundValue to walletService and getGasRefundValue to web3Service
+
+# 0.38.2
+
+- Add `getTokenIdFromTx` method to retrieve the tokenId from a transaction hash
+
 # 0.38.1
 
-- add `transferFrom` to walletService
+- Add `transferFrom` to walletService
 
 # 0.38.0